day06完成：根据分类ID查询菜品及其口味

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -29,7 +29,7 @@ public class DishController {
 
     @ApiOperation("根据分类id查询菜品")
     @GetMapping("list")
-    public Result<List<Dish>> getByCategoryId(Integer categoryId){
+    public Result<List<Dish>> getByCategoryId(Long categoryId){
 
         List<Dish> list =  dishService.getByCategoryId(categoryId);
