+++ sky-server/src/main/java/com/sky/mapper/DishMapper.java
@@ -52,5 +52,5 @@ public interface DishMapper {
 
 
     @Select("select * from dish where category_id = #{categoryId}")
-    List<Dish> getByCategoryId(Integer categoryId);
+    List<Dish> getByCategoryId(Long categoryId);
 }
